--- 7.30-Gameserver/GameUtils.h
@@ -1,14 +1,0 @@
-#pragma once
-
-namespace GameUtils
-{
-	namespace Siphon
-	{
-		void ApplySiphonEffectToEveryone();
-	}
-
-	namespace Snow
-	{
-		void SetSnow();
-	}
-}
