+++ 7.30-Gameserver/Hooks.h
@@ -1,53 +1,33 @@
 #pragma once
-#include "CustomSDK.hpp"
-using namespace SDK;
-
-extern enum ENetMode;
 
 namespace Hooks
 {
-    static void KickPlayerHook(AGameSession*, AController*) { return; }
+	static inline void (*KickPlayer)(AGameSession*, AController*);
+	static void KickPlayerHook(AGameSession*, AController*) { return; }
 
-    namespace hNet
-    {
-        inline __int64 (*Org_DispatchRequest)(__int64, __int64*, int);
-        __int64 Hook_DispatchRequest(__int64 a1, __int64* a2, int a3);
+	void StartServer()
+	{
+		CREATEHOOK(BaseAddress() + 0x2cbfc20, GetNetModeWorld, nullptr);
 
-        inline void (*Org_TickFlush)(UNetDriver*);
-        void Hook_TickFlush(UNetDriver* NetDriver);
+		CREATEHOOK(BaseAddress() + 0xa59280, DispatchRequestHook, &DispatchRequest);
 
-        ENetMode Hook_GetNetModeWorld();
-    }
+		CREATEHOOK(BaseAddress() + 0x28b0d20, KickPlayerHook, &KickPlayer);
 
-    namespace hGamemode
-    {
-        inline void (*Org_HandleStartingNewPlayer)(AFortGameModeAthena*, AFortPlayerControllerAthena*);
-        void Hook_HandleStartingNewPlayer(AFortGameModeAthena* GameMode, AFortPlayerControllerAthena* NewPlayer);
+		*(bool*)(BaseAddress() + 0x5a14019) = false; // GIsClient
 
-        inline bool (*Org_ReadyToStartMatch)(AFortGameModeAthena*);
-        bool Hook_ReadyToStartMatch(AFortGameModeAthena* GameMode);
+		((UFortEngine*)UFortEngine::GetEngine())->GameInstance->LocalPlayers.Remove(0);
 
-        APawn* Hook_SpawnDefaultPawnFor(AGameModeBase* GameMode, AController* NewPlayer, AActor* StartSpot);
-    }
+		UKismetSystemLibrary::GetDefaultObj()->ExecuteConsoleCommand(UWorld::GetWorld(), L"open Athena_Terrain", nullptr);
 
-    namespace hAbilities
-    {
-        void Hook_InternalServerTryActivateAbility(UAbilitySystemComponent* AbilitySystemComponent, FGameplayAbilitySpecHandle Handle, bool InputPressed, const FPredictionKey& PredictionKey, const FGameplayEventData* TriggerEventData);
-    }
+		MemoryUtils::NullFunction(BaseAddress() + 0x1014cc0); // changegamesession
+		MemoryUtils::NullFunction(MemoryUtils::SigScan("48 89 5C 24 ? 57 48 83 EC 30 48 8B 41 28 48 8B DA 48 8B F9 48 85 C0 74 34 48 8B 4B 08 48 8D")); // widget class
 
-    namespace hPlayer
-    {
-        inline void (*Org_ServerLoadingScreenDropped)(AFortPlayerControllerAthena*);
-        void Hook_ServerLoadingScreenDropped(AFortPlayerControllerAthena* Controller);
+		Player::HookAll();
+		Gamemode::HookAll();
+		Abilities::HookAll();
 
-        inline void (*Org_ServerAttemptInteract)(UObject* Context, FFrame* Stack, void* Ret);
-        void Hook_ServerAttemptInteract(UObject* Context, FFrame* Stack, void* Ret);
+		CREATEHOOK(BaseAddress() + 0x29ce460, TickFlushHook, &TickFlush);
 
-        inline void (*Org_ClientOnPawnDied)(AFortPlayerControllerAthena*, FFortPlayerDeathReport&);
-        void Hook_ClientOnPawnDied(AFortPlayerControllerAthena* DeadController, FFortPlayerDeathReport DeathReport);
-
-        void Hook_ServerAcknowlegePossession(APlayerController* Controller, APawn* Pawn);
-        void Hook_ServerExecuteInventoryItem(AFortPlayerControllerAthena* Controller, FGuid ItemGuid);
-        void Hook_ServerSendZiplineState(AFortPlayerPawn* Pawn, FZiplinePawnState InZiplineState);
-    }
+		MH_EnableHook(MH_ALL_HOOKS);
+	}
 }
