+++ 7.30-Gameserver/Net.h
@@ -1,6 +1,4 @@
 #pragma once
-#include "SDK/SDK.hpp"
-using namespace SDK;
 
 enum ENetMode
 {
@@ -11,13 +9,51 @@ enum ENetMode
     NM_MAX,
 };
 
-namespace Net
+ENetMode GetNetModeWorld()
 {
-    inline UNetDriver* (*CreateNetDriver)(UEngine* Engine, UWorld* InWorld, FName NetDriverDefinition);
-    inline bool (*InitListen)(UNetDriver* Driver, void* InNotify, FURL& LocalURL, bool bReuseAddressAndPort, FString& Error);
-    inline void* (*SetWorld)(UNetDriver* NetDriver, UWorld* World);
-    inline void (*ServerReplicateActors)(UReplicationDriver* ReplicationDriver); // Set in Hook_ReadyToStartMatch.
+    return ENetMode::NM_DedicatedServer;
+}
 
-    void SetupOffsets();
-    void SetupHooks();
+ENetMode GetNetModeActor()
+{
+    return ENetMode::NM_DedicatedServer;
+}
+
+inline UNetDriver* (*CreateNetDriver)(UEngine* Engine, UWorld* InWorld, FName NetDriverDefinition);
+inline bool (*InitListen)(UNetDriver* Driver, void* InNotify, FURL& LocalURL, bool bReuseAddressAndPort, FString& Error);
+inline void* (*SetWorld)(UNetDriver* NetDriver, UWorld* World);
+inline void (*ServerReplicateActors)(UReplicationDriver* ReplicationDriver);
+
+static __int64 (*DispatchRequest)(__int64, __int64*, int);
+
+static __int64 DispatchRequestHook(__int64 a1, __int64* a2, int a3)
+{
+    *(int*)(__int64(a2) + 0x28) = 3; // sets McpIsDedicatedServer
+
+    // 3 as a3 sets enabled, if we returned original a3, it would just set us as sparkle specialist lol
+    return DispatchRequest(a1, a2, 3);
+}
+
+void (*TickFlush)(UNetDriver*);
+void TickFlushHook(UNetDriver* NetDriver)
+{
+    if (auto ReplicationDriver = NetDriver->ReplicationDriver)
+        ServerReplicateActors(ReplicationDriver);
+
+    if (GetAsyncKeyState(VK_F5) & 1)
+    {
+        GameUtils::Snow::SetSnow();
+    }
+
+    if (GetAsyncKeyState(VK_F6) & 1)
+    {
+        GameUtils::Siphon::ApplySiphonEffectToEveryone();
+    }
+
+    if (GetAsyncKeyState(VK_F7) & 1)
+    {
+        UKismetSystemLibrary::GetDefaultObj()->ExecuteConsoleCommand(UWorld::GetWorld(), L"startaircraft", nullptr);
+    }
+
+    return TickFlush(NetDriver);
 }
