--- 7.30-Gameserver/Server.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace Server
-{
-	void StartServer();
-}
